+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 4.0 GUI and Plugin Refactoring
 
-- 2024-??-?? 4.0.57
+- 2024-03-25 4.0.57
     - fix check_refs in BoxcarFeature
     - support Expert mode in plugin fields
     - add Session Count to Measurement metadata